use the same date hook in query history as in saved aggregations
